# Simplify filestore.InlineReader